made comment its own component
