+++ public/styles/blogs/blog.scss
@@ -126,37 +126,6 @@
   width: 796px;
   margin-left: -2.5em;
 }
-.blogCommentContainer {
-  margin-top: 40px;
-  display: flex;
-  flex-direction: column;
-  justify-content: left;
-}
-.blogCommentForm {
-  margin-bottom: 30px;
-}
-.blogCommentForm textarea {
-  height: 3em;
-  width: 23em;
-  border: solid;
-}
-.blogCommentForm button {
-  margin-left: 1em;
-  background-color: black;
-  color: white;
-  border: 1px solid black;
-  cursor: pointer;
-}
-.blogCommentList {
-  display: flex;
-  flex-direction: column;
-  justify-content: left;
-}
-.blogSingleComment {
-  display: flex;
-  flex-direction: column;
-  margin-bottom: 30px;
-}
 
 //Mobile
 @media only screen and (max-width: 600px) {
@@ -203,9 +172,6 @@
     height: 200px;
     width: 355px;
     margin-left: 0;
-  }
-  .blogCommentContainer {
-    margin-left: 1em;
   }
 }
 
@@ -260,16 +226,6 @@
   .blogPostText {
     font-size: 1.7em;
   }
-  .blogCommentForm label {
-    font-size: 1.2em;
-  }
-  .blogCommentForm button {
-    width: 5em;
-    height: 2em;
-  }
-  .blogSingleComment p {
-    font-size: 1.4em;
-  }
 }
 
 //Desktop
@@ -308,22 +264,5 @@
   }
   .blogIframe {
     margin-left: 0;
-  }
-  .blogCommentForm label {
-    font-size: 1.2em;
-    margin-top: 1em;
-    margin-bottom: 0.5em;
-  }
-  .blogCommentForm button {
-    width: 4em;
-    height: 2em;
-    font-size: 1.4em;
-  }
-  .blogCommentForm textarea {
-    height: 4em;
-    width: 37em;
-  }
-  .blogSingleComment p {
-    font-size: 1.4em;
   }
 }
