+++ public/styles/index.scss
@@ -94,3 +94,4 @@ nav a {
 @import './mobileNavbar.scss';
 @import './navbar.scss';
 @import './submit.scss';
+@import './comments.scss';
